Stable function for giant gui program before changes. About to change function in a major way:
1: We are converting rotation to absolute instead of relative.
We don't have an encoder, but we'll set it to 0 at the beginning of the scan.
2: DefineScan will produce coordinates for all 4 axes not just X/Y
3: Gridscan will accept coordinates for all 4 axes. This means we are less
limited by scan types (we can do a partial rotation, for instance).
4: Gridscan will save leftover scan coordinates to a file for easy resuming.
My main intention is for "solving" the USB disconnect problem by restarting
 the whole pi and resuming
5: Upon booting, pi will automatically load this program, zero axes XYZ,
and call gridscan with scan coordinates IF a file is present.
6: upon successful completion we'll just delete or rename that file