A bunch of changes.
DefineScan now produces a dictionary with four keys X,Y,Z,R, each containing
a list of where that axis should be at every single point in the scan.
This also means that R scanning is considered absolute despite lack of endstop.

Gridscan now accepts two arguments: A dictionary of scan locations for all X,Y,Z,R, and a dictionary for other misc conditions.
Gridscan now capable of accepting arbitrary scan locations, as long as each axis
is a list of the same size. Partial rotations are thus possible.

When a scan is started the first time, default conditions are granted, user
is asked for file save location...

If a scan fails due to a USB disconnect now, the program will beep as before
waiting to see if the user will come and save the scan.

After a few seconds, however, the raspberry pi will restart after saving the
remaining scan to a pickle file.

The raspberry pi is set to load the 4steppergui automatically upon reboot:
How this is done is up to the user (it was a PITA to get to work in a not good way),
but what is set up on the original device is to autorun a bash file that opens
a terminal with the program running. Specifically, in /home/pi/.config/lxsession/LXDE-pi/autostart
add
@/home/pi/Desktop/bashfile.sh
to the last line.

The bashfile looks like:

lxterminal --command="/bin/bash -c '/home/pi/Desktop/ladybug/4steppergui.py; /bin/bash'"

and don't foret to make it and python file executable.
For whatever reason, this is the only thing that worked!!!

When the 4steppergui loads, it checks to see if there is a pickle file
in the folder location. If not, do nothing. If so, home the axes and run gridscan
with the saved parameters. Rinse and repeat...

A ridiculous bandaid solution to a ridiculous problem.